dedicated files for import schema tests. more cleanup
