+++ api/tests/test_blog_posts.py
@@ -5,11 +5,12 @@ from rest_framework.test import APITestCase
 from data.factories import BlogPostFactory, BlogTagFactory
 
 
-class TestBlogApi(APITestCase):
-    def test_get_published_blog_posts(self):
-        """
-        The API should only return published blog posts
-        """
+class BlogPostListApiTest(APITestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.url = reverse("blog_posts_list")
+
+    def test_can_list_blog_posts_if_published(self):
         published_blog_posts = [
             BlogPostFactory(published=True),
             BlogPostFactory(published=True),
@@ -19,13 +20,12 @@ class TestBlogApi(APITestCase):
             BlogPostFactory(published=False),
             BlogPostFactory(published=False),
         ]
-        response = self.client.get(reverse("blog_posts_list"))
+        response = self.client.get(self.url)
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         body = response.json()
-
         self.assertEqual(body.get("count"), 2)
-
-        results = body.get("results", [])
+        results = body["results"]
 
         for published_blog_post in published_blog_posts:
             self.assertTrue(any(x["id"] == published_blog_post.id for x in results))
@@ -33,26 +33,7 @@ class TestBlogApi(APITestCase):
         for draft_blog_post in draft_blog_posts:
             self.assertFalse(any(x["id"] == draft_blog_post.id for x in results))
 
-    def test_get_single_blog_post(self):
-        """
-        The API should return single blog posts if they are published
-        """
-        tag = BlogTagFactory(name="Test tag")
-        published_blog_post = BlogPostFactory(published=True)
-        published_blog_post.tags.add(tag)
-        draft_blog_post = BlogPostFactory(published=False)
-
-        response = self.client.get(reverse("single_blog_post", kwargs={"pk": published_blog_post.id}))
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertIn("Test tag", response.json()["tags"])
-
-        response = self.client.get(reverse("single_blog_post", kwargs={"pk": draft_blog_post.id}))
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-    def test_filter_blog_post(self):
-        """
-        The API should return filtered blog posts
-        """
+    def test_can_filter_blog_posts_by_tag(self):
         tag = BlogTagFactory(name="Test")
         other = BlogTagFactory()
         good_post = BlogPostFactory(published=True)
@@ -61,7 +42,8 @@ class TestBlogApi(APITestCase):
         post = BlogPostFactory(published=True)
         post.tags.add(other)
 
-        response = self.client.get(reverse("blog_posts_list"), {"tag": "Test"})
+        response = self.client.get(self.url, {"tag": "Test"})
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.json()["count"], 1)
         self.assertEqual(response.json()["results"][0]["id"], good_post.id)
@@ -79,11 +61,35 @@ class TestBlogApi(APITestCase):
         draft_blog_post = BlogPostFactory(published=False)
         draft_blog_post.tags.add(draft_tag)
 
-        response = self.client.get(reverse("blog_posts_list"))
+        response = self.client.get(self.url)
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         body = response.json()
-
         tags = body.get("tags", [])
         self.assertIn("Used", tags)
         self.assertNotIn("Unused", tags)
         self.assertNotIn("Draft", tags)
+
+
+class BlogPostDetailApiTest(APITestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.blog_post = BlogPostFactory(published=False)
+        cls.url = reverse("single_blog_post", kwargs={"pk": cls.blog_post.id})
+
+    def test_cannot_get_blog_post_if_not_published(self):
+        response = self.client.get(self.url)
+
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+    def test_can_get_blog_post_if_published(self):
+        tag = BlogTagFactory(name="Test tag")
+        self.blog_post.published = True
+        self.blog_post.save()
+        self.blog_post.tags.add(tag)
+
+        response = self.client.get(self.url)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        body = response.json()
+        self.assertIn("Test tag", body["tags"])
