+++ api/tests/test_initial_data.py
@@ -34,7 +34,7 @@ class InitialDataApiTest(APITestCase):
         cls.video_tutorial = VideoTutorialFactory(published=True)
         cls.url = reverse("initial_data")
 
-    def test_get_unauthenticated(self):
+    def test_can_get_unauthenticated(self):
         response = self.client.get(self.url)
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -63,7 +63,7 @@ class InitialDataApiTest(APITestCase):
         self.assertEqual(len(body["lineMinistries"]), len(Canteen.Ministries))
 
     @freeze_time(timezone.now() + datetime.timedelta(days=2))
-    def test_get_unauthenticated_no_upcoming_community_events(self):
+    def test_can_get_unauthenticated_no_upcoming_community_events(self):
         response = self.client.get(self.url)
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -73,7 +73,7 @@ class InitialDataApiTest(APITestCase):
         self.assertEqual(len(body["communityEvents"]), 0)
 
     @authenticate
-    def test_get_authenticated(self):
+    def test_can_get_authenticated(self):
         response = self.client.get(self.url)
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -94,7 +94,7 @@ class InitialDataApiTest(APITestCase):
         self.assertEqual(len(body["canteenPreviews"]), 0)
 
     @authenticate
-    def test_get_authenticated_and_manager(self):
+    def test_can_get_authenticated_and_manager(self):
         self.canteen.managers.set([authenticate.user])
 
         response = self.client.get(self.url)
