+++ api/tests/test_purchases_import.py
@@ -1,6 +1,4 @@
 import hashlib
-import json
-import re
 from datetime import date
 from decimal import Decimal
 from pathlib import Path
@@ -8,103 +6,28 @@ from unittest import skipIf
 from unittest.mock import patch
 
 from django.conf import settings
-from django.test import TestCase
 from django.test.utils import override_settings
 from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
 
 from api.tests.utils import assert_import_failure_created, authenticate
-from api.views.purchase_import import PURCHASE_ID_SCHEMA_FILE_PATH, PURCHASE_SIRET_SCHEMA_FILE_PATH
 from data.factories import CanteenFactory
 from data.models import ImportFailure, ImportType
 from data.models.creation_source import CreationSource
 from data.models.purchase import Purchase
 
 
-class PurchasesSchemaTest(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.schemas = {
-            "siret": json.load(open(PURCHASE_SIRET_SCHEMA_FILE_PATH)),
-            "id": json.load(open(PURCHASE_ID_SCHEMA_FILE_PATH)),
-        }
-
-    @staticmethod
-    def _pattern_for(schema, field_name):
-        field = next(f for f in schema["fields"] if f["name"] == field_name)
-        return field["constraints"]["pattern"]
-
-    def test_famille_produits_regex(self):
-        for schema_name, schema in self.schemas.items():
-            pattern = self._pattern_for(schema, "famille_produits")
-            for VALUE_OK in ["PRODUITS_LAITIERS", "PRODUITS_LAITIERS ", " PRODUITS_LAITIERS "]:
-                with self.subTest(schema=schema_name, VALUE=VALUE_OK):
-                    self.assertTrue(re.match(pattern, VALUE_OK))
-            for VALUE_NOT_OK in ["", "TEST", "PRODUITS_LAITIERS,", "PRODUITS_LAITIERS,VIANDES_VOLAILLES"]:
-                with self.subTest(schema=schema_name, VALUE=VALUE_NOT_OK):
-                    self.assertFalse(re.match(pattern, VALUE_NOT_OK))
-
-    def test_categories_egalim_regex(self):
-        for schema_name, schema in self.schemas.items():
-            pattern = self._pattern_for(schema, "categories_egalim")
-            for VALUE_OK in [
-                "BIO",
-                "BIO ",
-                "BIO,COMMERCE_EQUITABLE",
-                "BIO,COMMERCE_EQUITABLE ",
-                " BIO,COMMERCE_EQUITABLE ",
-                " BIO, COMMERCE_EQUITABLE ",
-                " BIO,      COMMERCE_EQUITABLE ",
-                "BIO,BIO",
-            ]:
-                with self.subTest(schema=schema_name, VALUE=VALUE_OK):
-                    self.assertTrue(re.match(pattern, VALUE_OK))
-            for VALUE_NOT_OK in ["", "TEST"]:
-                with self.subTest(schema=schema_name, VALUE=VALUE_NOT_OK):
-                    self.assertFalse(re.match(pattern, VALUE_NOT_OK))
-
-    def test_origine_regex(self):
-        for schema_name, schema in self.schemas.items():
-            pattern = self._pattern_for(schema, "origine")
-            for VALUE_OK in [
-                "EUROPE",
-                "FRANCE",
-                "FRANCE ",
-                " FRANCE ",
-            ]:
-                with self.subTest(schema=schema_name, VALUE=VALUE_OK):
-                    self.assertTrue(re.match(pattern, VALUE_OK))
-            for VALUE_NOT_OK in ["", "TEST", "FRANCE,", "FRANCE,EUROPE"]:
-                with self.subTest(schema=schema_name, VALUE=VALUE_NOT_OK):
-                    self.assertFalse(re.match(pattern, VALUE_NOT_OK))
-
-    def test_definition_local_regex(self):
-        for schema_name, schema in self.schemas.items():
-            pattern = self._pattern_for(schema, "definition_local")
-            for VALUE_OK in [
-                "PAT",
-                " PAT ",
-                "COMMUNE",
-                "DEPARTEMENT",
-                "DEPARTEMENT ",
-                " DEPARTEMENT ",
-                "REGION",
-                "KM",
-            ]:
-                with self.subTest(schema=schema_name, VALUE=VALUE_OK):
-                    self.assertTrue(re.match(pattern, VALUE_OK))
-            for VALUE_NOT_OK in ["", "TEST", "DEPARTEMENT,", "DEPARTEMENT,REGION"]:
-                with self.subTest(schema=schema_name, VALUE=VALUE_NOT_OK):
-                    self.assertFalse(re.match(pattern, VALUE_NOT_OK))
-
-
 @skipIf(settings.SKIP_TESTS_THAT_REQUIRE_INTERNET, "Skipping tests that require internet access")
 class PurchasesImportApiErrorTest(APITestCase):
-    def test_unauthenticated(self):
+    @classmethod
+    def setUpTestData(cls):
+        cls.url = reverse("purchases_import")
+
+    def test_cannot_import_if_unauthenticated(self):
         self.assertEqual(Purchase.objects.count(), 0)
 
-        response = self.client.post(reverse("purchases_import"), {"type": "siret"})
+        response = self.client.post(self.url, {"type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -120,7 +43,7 @@ class PurchasesImportApiErrorTest(APITestCase):
         # header missing
         file_path = "./api/tests/files/achats/purchases_bad_no_header.csv"
         with open(file_path, "rb") as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -135,7 +58,7 @@ class PurchasesImportApiErrorTest(APITestCase):
         # wrong header
         file_path = "./api/tests/files/achats/purchases_bad_wrong_header.csv"
         with open(file_path, "rb") as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -150,7 +73,7 @@ class PurchasesImportApiErrorTest(APITestCase):
         # partial header
         file_path = "./api/tests/files/achats/purchases_bad_partial_header.csv"
         with open(file_path, "rb") as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -178,7 +101,7 @@ class PurchasesImportApiErrorTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_bad_wrong_header_typo.csv"
         with open(file_path, "rb") as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -254,7 +177,7 @@ class PurchasesImportApiErrorTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_bad_extra_columns.csv"
         with open(file_path, "rb") as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -281,7 +204,7 @@ class PurchasesImportApiErrorTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_bad_empty_rows.csv"
         with open(file_path) as canteen_file:
-            response = self.client.post(reverse("purchases_import"), {"file": canteen_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": canteen_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -306,7 +229,7 @@ class PurchasesImportApiErrorTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_bad.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -351,7 +274,7 @@ class PurchasesImportApiErrorTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_bad_one_error.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -368,7 +291,7 @@ class PurchasesImportApiErrorTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -387,7 +310,7 @@ class PurchasesImportApiErrorTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -409,7 +332,7 @@ class PurchasesImportApiErrorTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -430,7 +353,7 @@ class PurchasesImportApiErrorTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_bad_one_error.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -448,7 +371,7 @@ class PurchasesImportApiErrorTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_bad_corrupt.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -468,6 +391,10 @@ class PurchasesImportApiErrorTest(APITestCase):
 
 @skipIf(settings.SKIP_TESTS_THAT_REQUIRE_INTERNET, "Skipping tests that require internet access")
 class PurchasesImportApiSuccessTest(APITestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.url = reverse("purchases_import")
+
     @authenticate
     def test_import_good_purchases(self):
         """
@@ -478,7 +405,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 8)
@@ -540,7 +467,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good_with_empty_columns.xlsx"
         with open(file_path, "rb") as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 2)
@@ -553,7 +480,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good_separator_comma.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 1)
@@ -573,7 +500,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good_decimal_comma.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 1)
@@ -596,7 +523,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good.xlsx"
         with open(file_path, "rb") as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 1)
@@ -616,7 +543,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good_decimal_comma.xlsx"
         with open(file_path, "rb") as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 1)
@@ -640,7 +567,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
         # comma
         file_path = "./api/tests/files/achats/purchases_good_separator_comma.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 1)
@@ -649,7 +576,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
         # tab
         file_path = "./api/tests/files/achats/purchases_good_separator_tab.tsv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 1 + 1)
@@ -658,7 +585,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
         # semicolon
         file_path = "./api/tests/files/achats/purchases_good_separator_semicolon.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 2 + 1)
@@ -675,7 +602,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good.csv"
         with open(file_path) as purchase_file:
-            _ = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            _ = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(_process_chunk_mock.call_count, 8)
 
@@ -690,7 +617,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
         # first upload: success
         file_path = "./api/tests/files/achats/purchases_good.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 8)
@@ -698,7 +625,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         # second upload: duplicate warning
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 8)  # no additional purchases created
@@ -721,7 +648,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good_floating_number.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 1)
@@ -738,7 +665,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good_local_vs_circuit_court.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 2)
@@ -768,7 +695,7 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good_siret_caracteristics.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file, "type": "siret"})
+            response = self.client.post(self.url, {"file": purchase_file, "type": "siret"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 2)
@@ -801,13 +728,17 @@ class PurchasesImportApiSuccessTest(APITestCase):
 
 @skipIf(settings.SKIP_TESTS_THAT_REQUIRE_INTERNET, "Skipping tests that require internet access")
 class PurchasesImportIdApiErrorTest(APITestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.url = reverse("purchases_import")
+
     @authenticate
     def test_canteen_not_found_with_id(self):
         self.assertEqual(Purchase.objects.count(), 0)
 
         file_path = "./api/tests/files/achats/purchases_good_id.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file})
+            response = self.client.post(self.url, {"file": purchase_file})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 0)
@@ -820,6 +751,10 @@ class PurchasesImportIdApiErrorTest(APITestCase):
 
 @skipIf(settings.SKIP_TESTS_THAT_REQUIRE_INTERNET, "Skipping tests that require internet access")
 class PurchasesImportIdApiSuccessTest(APITestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.url = reverse("purchases_import")
+
     @authenticate
     def test_import_default_type_is_id(self):
         """
@@ -830,7 +765,7 @@ class PurchasesImportIdApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good_id.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file})
+            response = self.client.post(self.url, {"file": purchase_file})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 1)
@@ -850,7 +785,7 @@ class PurchasesImportIdApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good_id.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file})
+            response = self.client.post(self.url, {"file": purchase_file})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 1)
@@ -878,7 +813,7 @@ class PurchasesImportIdApiSuccessTest(APITestCase):
 
         file_path = "./api/tests/files/achats/purchases_good_id_caracteristics.csv"
         with open(file_path) as purchase_file:
-            response = self.client.post(reverse("purchases_import"), {"file": purchase_file})
+            response = self.client.post(self.url, {"file": purchase_file})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Purchase.objects.count(), 2)
