+++ api/tests/test_teledeclaration.py
@@ -4,14 +4,24 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 
 
-class TestTeledeclarationCampaignDatesApi(APITestCase):
-    def test_campaign_dates_list(self):
-        response = self.client.get(reverse("list_teledeclaration_campaign_dates"))
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
+class TeledeclarationCampaignDatesListApiTest(APITestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.url = reverse("list_teledeclaration_campaign_dates")
 
+    def test_campaign_dates_list(self):
+        response = self.client.get(self.url)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
         body = response.json()
         self.assertEqual(len(body), 5)
         self.assertEqual(body[-1]["year"], 2025)
+
+
+class TeledeclarationCampaignDatesRetrieveApiTest(APITestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.url = reverse("retrieve_teledeclaration_campaign_dates", kwargs={"year": 2024})
 
     def test_campaign_dates_retrieve(self):
         for date_freeze in [
@@ -21,11 +31,9 @@ class TestTeledeclarationCampaignDatesApi(APITestCase):
         ]:
             with self.subTest(DATE=date_freeze["date"]):
                 with freeze_time(date_freeze["date"]):
-                    response = self.client.get(
-                        reverse("retrieve_teledeclaration_campaign_dates", kwargs={"year": 2024})
-                    )
-                    self.assertEqual(response.status_code, status.HTTP_200_OK)
+                    response = self.client.get(self.url)
 
+                    self.assertEqual(response.status_code, status.HTTP_200_OK)
                     body = response.json()
                     self.assertEqual(body["year"], 2024)
                     self.assertEqual(body["inTeledeclaration"], date_freeze["in_teledeclaration"])
